Move MEF exports to the real classes, and remove indirection
